Added the Ability to Restart the Server
This is a fork of the MinecraftStartupScript made by ForsakenDev. This change allows your server to automatically restart in the event of a crash. To enable this functionality, add a line you your crontab to automatically call this this script and pass in the "serverCheck" argument (set this to run every 2 minutes or so). The script will check if the server is down and and will automatically start it back up.

NOTE: If you wish to keep your server down for a long period of time (for example, when updating your Minecraft server version) you must comment out the serverCheck line in your crontab otherwise the script will start your server back up while you are working on it.